+++ src/uactor/examples/dependency_injection.rs
@@ -1,5 +1,5 @@
 use time::ext::NumericalStdDuration;
-use uactor::actor::MessageSender;
+use uactor::actor::abstract_actor::MessageSender;
 
 use uactor::system::System;
 
@@ -13,7 +13,7 @@ use crate::messages::{MessageWithoutReply, PingMsg, PongMsg};
 use crate::services::{Service1, Service2};
 
 mod messages {
-    use uactor::message::{Message, Reply};
+    use uactor::actor::message::{Message, Reply};
 
     pub struct PingMsg(pub Reply<PongMsg>);
 
@@ -32,10 +32,10 @@ mod messages {
 mod actor1 {
     use tokio::sync::mpsc::UnboundedSender;
 
-    use uactor::actor::{Actor, HandleResult, Handler, MessageSender};
-    use uactor::context::extensions::Service;
-    use uactor::context::Context;
-    use uactor::di::{Inject, InjectError};
+    use uactor::actor::abstract_actor::{Actor, HandleResult, Handler, MessageSender};
+    use uactor::actor::context::extensions::Service;
+    use uactor::actor::context::Context;
+    use uactor::dependency_injection::{Inject, InjectError};
     use uactor::system::System;
 
     use crate::actor2::{Actor2Msg, Actor2Ref};
@@ -104,10 +104,10 @@ mod actor1 {
 }
 
 mod actor2 {
-    use uactor::actor::{Actor, HandleResult, Handler};
-    use uactor::context::extensions::Service;
-    use uactor::context::Context;
-    use uactor::di::{Inject, InjectError};
+    use uactor::actor::abstract_actor::{Actor, HandleResult, Handler};
+    use uactor::actor::context::extensions::Service;
+    use uactor::actor::context::Context;
+    use uactor::dependency_injection::{Inject, InjectError};
     use uactor::system::System;
 
     use crate::messages::{PingMsg, PongMsg, PrintMessage};
