+++ src/uactor/examples/single_channel_actor.rs
@@ -1,4 +1,4 @@
-use uactor::actor::MessageSender;
+use uactor::actor::abstract_actor::MessageSender;
 use uactor::system::System;
 
 use crate::actor1::Actor1;
@@ -7,7 +7,7 @@ use crate::actor1::Actor1Ref;
 use crate::messages::PingMsg;
 
 mod messages {
-    use uactor::message::{Message, Reply};
+    use uactor::actor::message::{Message, Reply};
 
     pub struct PingMsg(pub Reply<PongMsg>);
 
@@ -18,8 +18,8 @@ mod messages {
 }
 
 mod actor1 {
-    use uactor::actor::{Actor, HandleResult, Handler};
-    use uactor::context::Context;
+    use uactor::actor::abstract_actor::{Actor, HandleResult, Handler};
+    use uactor::actor::context::Context;
 
     use crate::messages::{PingMsg, PongMsg};
