+++ src/uactor/src/lib.rs
@@ -1,10 +1,7 @@
-pub mod actor;
-pub mod context;
-pub mod data_publisher;
-pub mod datasource;
-pub mod di;
-pub mod message;
-pub mod select;
+pub mod dependency_injection;
 pub mod system;
 
 pub use paste::paste;
+
+pub mod data;
+pub mod actor;
